+++ ext/manifest.js
@@ -3,7 +3,7 @@ module.exports = {
   description: 'Examine and edit proxied requests',
   version: '0.0.2',
   manifest_version: 3,
-  permissions: ['scripting', 'proxy', 'debugger'],
+  permissions: ['scripting', 'proxy', 'debugger', 'notifications'],
   devtools_page: 'devtool.html',
   host_permissions: ['<all_urls>'],
   web_accessible_resources: [