+++ src/config.mjs
@@ -2,11 +2,17 @@ import fs from 'fs';
 import path from 'path';
 import deepmerge from 'deepmerge';
 import getUserConfigFile from './read-user-config.js';
-import {
-  PROJECT_ROOT,
-  CONFIG_FILE_NAME,
-  CACHE_DIR_NAME,
-} from './constants.mjs';
+import { CONFIG_FILE_NAME, CACHE_DIR_NAME } from './constants.mjs';
+
+const prepCacheDir = (cwd) => {
+  const cacheDir = path.join(cwd, CACHE_DIR_NAME);
+  if (fs.existsSync(cacheDir)) {
+    return;
+  }
+
+  console.log(`Couldn't locale ${CACHE_DIR_NAME}/, creating one.`);
+  fs.mkdirSync(cacheDir);
+};
 
 export default function config(overrides = {}, cwd = process.cwd()) {
   const filepath = path.join(cwd, CONFIG_FILE_NAME);
@@ -16,11 +22,13 @@ export default function config(overrides = {}, cwd = process.cwd()) {
 
   const logName = `sever.${new Date().toISOString().split('T')[0]}.log`;
 
+  prepCacheDir(cwd);
+
   const config = deepmerge(
     {
       cwd,
       filepath,
-      logLocation: path.join(PROJECT_ROOT, logName),
+      logLocation: path.join(cacheDir, logName),
       forward: {},
       noProxy: ['<-loopback->'],
       ...userConfig,
@@ -28,10 +36,7 @@ export default function config(overrides = {}, cwd = process.cwd()) {
     overrides
   );
 
-  if (fs.existsSync(cacheDir)) {
-    config.cache = deepmerge(config.cache || {}, { dir: cacheDir });
-    config.logLocation = path.join(cacheDir, logName);
-  }
+  config.cache = deepmerge(config.cache || {}, { dir: cacheDir });
 
   return config;
 }