+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.10 Better image handling, auto mocking
+
+- Let users specify a filepath during an auto-mock. This filepath will be read
+  from disk when an auto-mock is used.
+- Improve `nextjs-graphql-example`. Add a demo of an image placeholder.
+- Add Chrome Extension notifications. Config updates fire notifications.
+- Add `ping` & `shutdown` sub-commmands to `jam-server`
+- Add `tail` sub-command for tailing a current log
+- Create a `.jambox` folder if one doesn't exist in cwd
+
 ## 0.0.9 NextJS demo, better watch mode
 
 - A tidy GraphQL example